fixed bug in hp

+++ wow_rl_api/wow_rl_api/Environment.h
@@ -13,6 +13,6 @@ public:
 
 private:
 
-	float GetRemainingHp(int hp);
+	float GetRemainingHp(float hp);
 };
